fix(continente-api): fix field type

+++ apps/external-apis/continente-api/app/Domain/Entity/ProductEntity.php
@@ -8,7 +8,7 @@ class ProductEntity
         private string $externalId,
         private string $name,
         private ?string $category,
-        private int $priceCents,
+        private float $priceCents,
         private string $currency,
         private bool $available
     ) {}
